Minimum viable pyproject.toml for PEP-518 compliance  #2
The new pyproject.toml specifies the minimum requirements necessary to build
the project's wheel. This makes this project up to the PEP-518 standard.

Specifically, this change adds torch as a build-time dependency. The
version of torch can be set by the installer when they specify a version
of torch in their own project settings (e.g. a pyproject.toml,
requirements.txt, setup.py, environment.yaml, etc.). Alteratively,
if torch is in the current installing environment, that version will be
used. If not specified, the latest torch version will be used.

This change preserves using `setup.py` for project building.
To ensure that the existing installation semantics are preserved,
the `build-backend` is also specified in `pyproject.toml`.