Set torch build-time dependency
Sets `torch` as a build-time dependency in `pyproject.toml` in `requires`
under the `build-system` section.

NOTE: Due to difficulties in configuring this build-time dependency to
respect what users have in their requirements (e.g. torch 1.4.0, 1.5.1,
or 1.6.0) and my need to have this working for 1.5.1, this dep is set to
be < 1.6.0. (TODO: enable users to override this dependency).